Fix circular dependency in src/datatypes.ts
