+++ src/rdf.internal.ts
@@ -25,7 +25,7 @@ import rdfJsDatasetModule from "@rdfjs/dataset";
 export const rdfJsDataset = rdfJsDatasetModule.dataset;
 import * as RdfJs from "@rdfjs/types";
 import { IriString } from "./interfaces";
-import type { XmlSchemaTypeIri } from "./datatypes";
+import type { XmlSchemaTypeIri } from "./xmlSchemaTypes";
 
 export const localNodeSkolemPrefix =
   "https://inrupt.com/.well-known/sdk-local-node/" as const;
